Feature/create appointment rest service

+++ backend/src/main/java/com/krol/nail/salon/dtos/ServiceDto.java
@@ -1,4 +1,4 @@
 package com.krol.nail.salon.dtos;
 
-public record ServiceDto(String category, String name, String description, String price, double duration, boolean popular) {
+public record ServiceDto(String category, String name, String description, String price, Long duration, boolean popular) {
 }
